+++ src/header/Stash Tag Image Cropper.user.js
@@ -1,0 +1,15 @@
+// ==UserScript==
+// @name        Stash Tag Image Cropper
+// @namespace   %NAMESPACE%
+// @description Adds an image cropper to tag page
+// @version     0.1.0
+// @author      7dJx1qP
+// @match       %MATCHURL%
+// @resource    IMPORTED_CSS https://raw.githubusercontent.com/fengyuanchen/cropperjs/main/dist/cropper.min.css
+// @grant       unsafeWindow
+// @grant       GM_getResourceText
+// @grant       GM_addStyle
+// @require     %LIBRARYPATH%
+// @require     https://raw.githubusercontent.com/fengyuanchen/cropperjs/main/dist/cropper.min.js
+// @require     %FILEPATH%
+// ==/UserScript==